Rename functions/GravityForms to functions/gravityForms
